+++ helloservice/HelloService.cpp
@@ -33,6 +33,6 @@ String8 HelloService::sayhello_to(String8 name) {
     // do nothing here. Processing in function onTransact()
 }
 
-uint8_t HelloService::test(uint8_t val) {
+int32_t HelloService::test(int32_t val) {
     
 }
